Update CMakeLists to handle namespaces in binary names.
Usually target_link_libraries handles this, but using normal cmake targets and functions will result in the normal c/++ compiler being used instead of mex; this means we need to handle namespace striping ourselves.

The regex could probably be improved. I'm not sure what characters are used to seperate entries if multiple entries are present in MOOS_LIBRARIES . There should only be one right now and it probably won't change, but if more are added in future this will probably break. The regex should be updated to match all things that aren't the seperator followed by :: . CMakee regexes are rather limited and don't have character classes that neatly handle things like whitespace or not-whitespace for eaxmple.